requirements.txt
streamlitstreamlit
matplotlib
pandas
numpy